refactor: decomposition SignUp
