+++ app/[lang]/signup/page.tsx
@@ -10,74 +10,12 @@ import FormControl from '@mui/material/FormControl';
 import Link from '@mui/material/Link';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
-import Stack from '@mui/material/Stack';
-import MuiCard from '@mui/material/Card';
-import { styled } from '@mui/material/styles';
 import { useCreateUserWithEmailAndPassword } from 'react-firebase-hooks/auth';
 import { auth } from '@/app/firebase/config';
+import { Card, SignUpContainer } from './styled';
 import z from 'zod';
-
-const Card = styled(MuiCard)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignSelf: 'center',
-  width: '100%',
-  padding: theme.spacing(4),
-  gap: theme.spacing(2),
-  margin: 'auto',
-  boxShadow:
-    'hsla(220, 30%, 5%, 0.05) 0px 5px 15px 0px, hsla(220, 25%, 10%, 0.05) 0px 15px 35px -5px',
-  [theme.breakpoints.up('sm')]: {
-    width: '450px',
-  },
-  ...theme.applyStyles('dark', {
-    boxShadow:
-      'hsla(220, 30%, 5%, 0.5) 0px 5px 15px 0px, hsla(220, 25%, 10%, 0.08) 0px 15px 35px -5px',
-  }),
-}));
-
-const SignUpContainer = styled(Stack)(({ theme }) => ({
-  height: 'calc((1 - var(--template-frame-height, 0)) * 100dvh)',
-  minHeight: '100%',
-  padding: theme.spacing(2),
-  [theme.breakpoints.up('sm')]: {
-    padding: theme.spacing(4),
-  },
-  '&::before': {
-    content: '""',
-    display: 'block',
-    position: 'absolute',
-    zIndex: -1,
-    inset: 0,
-    backgroundImage:
-      'radial-gradient(ellipse at 50% 50%, hsl(210, 100%, 97%), hsl(0, 0%, 100%))',
-    backgroundRepeat: 'no-repeat',
-    ...theme.applyStyles('dark', {
-      backgroundImage:
-        'radial-gradient(at 50% 50%, hsla(210, 100%, 16%, 0.5), hsl(220, 30%, 5%))',
-    }),
-  },
-}));
-
-export const formSchema = z.object({
-  email: z.email({ error: 'Enter valid email' }),
-  password: z
-    .string()
-    .min(8, { error: 'Must be at least 8 characters' })
-    .refine((value) => /[a-z]/.test(value), {
-      error: 'Add one lowercase',
-    })
-    .refine((value) => /[A-Z]/.test(value), {
-      error: 'Add one uppercase',
-    })
-    .refine((value) => /\d/.test(value), {
-      error: 'Add one number',
-    })
-    .refine((value) => /[!@#$%^&*+-]/.test(value), {
-      error: 'Add one special character ',
-    }),
-});
-type FormData = z.infer<typeof formSchema>;
+import { formSchema, type FormData } from './types';
+import { useRouter } from 'next/navigation';
 
 const initialFormState = {
   email: '',
@@ -85,13 +23,15 @@ const initialFormState = {
 };
 
 export default function SignUp(props: { disableCustomTheme?: boolean }) {
-  const [userFormData, setFormData] = useState<Partial<FormData>>({
+  const [userFormData, setFormData] = useState<FormData>({
     email: '',
     password: '',
   });
   const [showErrors, setShowErrors] = useState(false);
 
   const [createUser] = useCreateUserWithEmailAndPassword(auth);
+
+  const router = useRouter();
 
   const formData = {
     ...initialFormState,
@@ -123,8 +63,9 @@ export default function SignUp(props: { disableCustomTheme?: boolean }) {
         userFormData.email,
         userFormData.password
       );
-      reset();
       console.log({ response });
+      reset();
+      router.push('/');
     } catch (error) {
       console.log(error);
     }
